dev-MetOfficeAgent-withinStack: consolidated fixtures in conftest.py
